Convert autocmd macros to new `auc-` macro
The old macro only allowed autocmds where the right hand side is a
string. The new macro allows the righthand side to be either a string or
a callback. Because of this, it is also possible to remove `global`
keywords on functions that only serve as callbacks to autocmds. All such
`global` keywords have been removed in this commit.